+++ template/app/(main)/screen1/index.tsx
@@ -1,4 +1,5 @@
-import { Text, ThemedView } from "@/components/atoms";
+import { Text } from "@/components/atoms";
+import MainScreenWrapper from "@/components/templates/MainScreenWrapper";
 import React from "react";
 import { Appearance, Switch, useColorScheme } from "react-native";
 
@@ -6,7 +7,7 @@ export default function Screen1() {
   const colorScheme = useColorScheme();
 
   return (
-    <ThemedView style={{ flex: 1, padding: 16 }}>
+    <MainScreenWrapper>
       <Text>Screen1</Text>
       <Switch
         value={colorScheme == "dark"}
@@ -14,6 +15,6 @@ export default function Screen1() {
           Appearance.setColorScheme(colorScheme == "dark" ? "light" : "dark");
         }}
       />
-    </ThemedView>
+    </MainScreenWrapper>
   );
 }