+++ template/components/molecules/common/index.tsx
@@ -1,3 +1,4 @@
 export { default as Collapsible } from "./Collapsible";
 export { default as FormInput } from "./FormInput";
 export { default as Checkbox } from "./Checkbox";
+export { default as NoResults } from "./NoResults";