+++ template/components/wrappers/Card/styles.ts
@@ -1,0 +1,21 @@
+import { COLORS } from "@/constants/Colors";
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    backgroundColor: COLORS.light.white,
+    shadowColor: COLORS.light.black,
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+    elevation: 5,
+    padding: 16,
+    borderRadius: 16,
+    width: "100%",
+  },
+});
+
+export default styles;